+++ package.json
@@ -13,7 +13,7 @@
     "build": "webpack --config webpack.config.js",
     "clean": "rm -rf ./dist",
     "~build": "tsc --watch",
-    "test": "TS_NODE_FILES=true mocha -r ts-node/register test/**/*.test.ts",
+    "test": "TS_NODE_FILES=true mocha -r ts-node/register test/**/*.test.ts src/lib/CLValue2/*.test.ts",
     "test:coverage": "nyc npm run test",
     "docs": "typedoc --out docs/code/ src",
     "publish-doc": "yarn run docs && gh-pages -d docs -r https://github.com/casper-ecosystem/casper-client-sdk-docs",
