+++ src/lib/CLValue2/Abstract.ts
@@ -1,0 +1,8 @@
+export abstract class CLType {
+  abstract toString(): string;
+}
+
+export abstract class CLValue {
+  abstract clType(): CLType;
+  abstract value(): any;
+}
