+++ src/lib/CLValue2/List.ts
@@ -1,0 +1,80 @@
+import { CLType, CLValue } from "./Abstract"; 
+
+export class ListType<T extends CLType> extends CLType {
+  inner: T;
+  constructor(inner: T) {
+    super();
+    this.inner = inner;
+  }
+
+  toString(): string {
+    return `List (${this.inner.toString()})`;
+  }
+}
+
+export class List<T extends CLValue> extends CLValue {
+  v: Array<T>;
+  vectorType: CLType;
+
+  constructor(v: Array<T> | CLType) {
+    super();
+    if (Array.isArray(v) && v[0].clType) {
+      const refType = v[0].clType();
+      if (
+        v.every(i => {
+          return i.clType().toString() === refType.toString();
+        })
+      ) {
+        this.v = v;
+        this.vectorType = refType;
+      } else {
+        throw Error('Invalid data provided.');
+      }
+    } else if (v instanceof CLType) {
+      this.vectorType = v;
+      this.v = [];
+    } else {
+      throw Error('Invalid data type(s) provided.');
+    }
+  }
+
+  value(): Array<T> {
+    return this.v;
+  }
+
+  clType(): CLType {
+    return new ListType(this.vectorType);
+  }
+
+  get(index: number): T {
+    return this.v[index];
+  }
+
+  set(index: number, item: T): void {
+    if (index >= this.v.length) {
+      throw new Error("Array index out of bounds.");
+    }
+    this.v[index] = item;
+  }
+
+  push(item: T): void {
+    if (item.clType().toString() === this.vectorType.toString()) {
+      this.v.push(item);
+    } else {
+      throw Error(`Incosnsistent data type, use ${this.vectorType.toString()}.`);
+    }
+  }
+
+  remove(index: number): void {
+    this.v.splice(index, 1);
+  }
+
+  // TBD: we can throw an error here, but returing undefined from empty list is typical JS behavior
+  pop(): T | undefined {
+    return this.v.pop();
+  }
+
+  size(): number {
+    return this.v.length;
+  }
+}
